Update README.md
We created a database using SQLite at a basic level for a supermarket. We took the product name and quantity for the product. We added the products to the database, deleted the products that we did not want in the database, and made updates to the database. We learned which products were available using the database. Our goal was to create a nice application while learning SQLite database.